+++ src-tauri/src/lib.rs
@@ -30,6 +30,8 @@ pub fn run() {
         .plugin(tauri_plugin_opener::init())
         .plugin(tauri_plugin_process::init())
         .plugin(tauri_plugin_updater::Builder::new().build())
+        .plugin(tauri_plugin_dialog::init())
+        .plugin(tauri_plugin_persisted_scope::init())
         .manage(ClickerState {
             running: Arc::new(AtomicBool::new(false)),
             run_generation: AtomicU64::new(0),